Extract light sessions view
